Backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
pdfplumber>=0.10.4
chromadb>=0.4.24
google-generativeai>=0.4.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-multipart>=0.0.9
sqlalchemy>=2.0.28
aiosqlite>=0.20.0
langchain-text-splitters>=0.0.1
greenlet>=3.0.0
openai>=1.12.0
